backend/requirements.txt
django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
django-filter==24.3
pillow==10.4.0
pytz==2024.1
gunicorn==20.1.0
psycopg2-binary==2.9.9